+++ lib/Alembic/Abc/Tests/PropertyTests.cpp
@@ -57,9 +57,9 @@ using Alembic::Util::uint32_t;
 
 void writeSimpleProperties(const std::string &archiveName)
 {
-    const unsigned int numChildren = 3;
+    const int numChildren = 3;
 
-    const unsigned int numSamples = 5;
+    const int numSamples = 5;
     const chrono_t dt = 1.0 / 24.0;
 
     TimeSampling ts(dt, 666.0); // uniform with cycle=dt starting at 666.0
@@ -125,13 +125,10 @@ void readSimpleProperties(const std::string &archiveName)
     IObject archiveTop = archive.getTop();
 
     // Determine the number of (top level) children the archive has
-    const unsigned int numChildren = archiveTop.getNumChildren();
+    const int numChildren = archiveTop.getNumChildren();
     TESTING_ASSERT( numChildren == 4 );
     std::cout << "The archive has " << numChildren << " children:"
               << std::endl;
-
-
-
 
     // Iterate through them, print out their names
     for (int ii=0; ii<numChildren; ii++)
@@ -144,7 +141,7 @@ void readSimpleProperties(const std::string &archiveName)
 
         // Properties
         ICompoundProperty props = child.getProperties();
-        size_t numProperties = props.getNumProperties();
+        int numProperties = props.getNumProperties();
 
         std::cout << "  ..and " << numProperties << " simple properties"
                   << std::endl;
@@ -237,13 +234,12 @@ void readSimpleProperties(const std::string &archiveName)
                 GetCompoundPropertyReaderPtr(props)->
                 getScalarProperty( propNames[jj] )->getTimeSampling();
 
-            size_t numSamples = ts->getNumStoredTimes();
+            int numSamples = ts->getNumStoredTimes();
 
 
             std::cout << "    ..and "
                       << ts->getTimeSamplingType() << std::endl
                       << "    ..and " << numSamples << " samples at times: ";
-
 
             if (numSamples > 0)
             {
@@ -368,7 +364,7 @@ void readSimpleProperties(const std::string &archiveName)
 
 void writeEmptyCompoundProperties(const std::string &archiveName)
 {
-    const unsigned int numChildren = 2;
+    const int numChildren = 2;
 
     // Create an archive for writing. Indicate that we want Alembic to
     //   throw exceptions on errors.
@@ -408,7 +404,7 @@ void readEmptyCompoundProperties(const std::string &archiveName)
     IObject archiveTop = archive.getTop();
 
     // Determine the number of (top level) children the archive has
-    const unsigned int numChildren = archiveTop.getNumChildren();
+    const int numChildren = archiveTop.getNumChildren();
     ABCA_ASSERT( numChildren == 2, "Wrong number of children (expected 2)");
     std::cout << "The archive has " << numChildren << " children:"
               << std::endl;
@@ -424,7 +420,7 @@ void readEmptyCompoundProperties(const std::string &archiveName)
 
         // Properties
         ICompoundProperty props = child.getProperties();
-        size_t numProperties = props.getNumProperties();
+        int numProperties = props.getNumProperties();
 
         std::cout << "  ..and " << numProperties << " properties"
                   << std::endl;
